Switching to SSV encoding and chancing cycle detection
Now using a variant of the SSV encoding instead of DITT. This makes it
a bit easier to control which fanins are enabled out-of-solver as well
as to break some symmetries, which should make enumeration more
efficient since it will find fewer duplicates. Similarly, cycle
detection is now done partially out-of-solver, which also makes
symmetry breaking easier.